refactor(bootstrap): extract resolution to RequirementResolver
This commit moves PyPI and graph-based resolution logic from Bootstrapper to new
  RequirementResolver class in requirement_resolver.py module. This
  separates resolution concerns from orchestration

Key changes:
  - Created RequirementResolver with resolve_source() and resolve_prebuilt()
  - Extracted _resolve_from_graph() and _resolve_from_version_source()
  - Git URL resolution stays in Bootstrapper (requires build orchestration)
  - Session state (_resolved_requirements, why stack) stays in Bootstrapper
  - Updated tests to use new resolver interface

This commit is phase 1 of the plan which was
generated using claude and can be found here
in the gist: https://gist.github.com/rd4398/35bb4d56b0c11b4b5eb5f194a0886e26
